requirements.txt
azure-functions
scikit-learn
requests